Addressed review comments (part 1)
