+++ cli/exceptions.py
@@ -61,14 +61,14 @@ class IRNotImplemented(IRException):
     pass
 
 
-class IRUnknownApplicationException(IRException):
+class IRUnknownSpecException(IRException):
     """
     This exceptions is raised when unknown application is
      started by user.
     """
     def __init__(self, app_name):
         self.app_name = app_name
-        super(IRUnknownApplicationException, self).__init__(
+        super(IRUnknownSpecException, self).__init__(
             "Application is unknown: '{}'".format(app_name))
